+++ src/extractors/config.rs
@@ -1,6 +1,5 @@
-use headers::www_authenticate::Challenge;
-
 use super::AuthenticationError;
+use crate::headers::www_authenticate::Challenge;
 
 pub trait ExtractorConfig {
     type Inner: Challenge;
@@ -8,7 +7,10 @@ pub trait ExtractorConfig {
     fn into_inner(self) -> Self::Inner;
 }
 
-impl<T> From<T> for AuthenticationError<<T as ExtractorConfig>::Inner> where T: ExtractorConfig {
+impl<T> From<T> for AuthenticationError<<T as ExtractorConfig>::Inner>
+where
+    T: ExtractorConfig,
+{
     fn from(config: T) -> Self {
         AuthenticationError::new(config.into_inner())
     }