+++ src/extractors/mod.rs
@@ -1,6 +1,8 @@
-mod errors;
-mod config;
+//! Type-safe authentication information extractors
+
 pub mod basic;
 pub mod bearer;
+mod config;
+mod errors;
 
 pub use self::errors::AuthenticationError;