+++ src/headers/www_authenticate/challenge/bearer/mod.rs
@@ -1,0 +1,12 @@
+//! Challenge for the "Bearer" HTTP Authentication Scheme
+
+mod builder;
+mod challenge;
+mod errors;
+
+pub use self::builder::BearerBuilder;
+pub use self::challenge::Bearer;
+pub use self::errors::Error;
+
+#[cfg(test)]
+mod tests;