+++ src/headers/www_authenticate/challenge/mod.rs
@@ -1,12 +1,13 @@
 use std::fmt::{Debug, Display};
 
-use bytes::Bytes;
 use actix_web::http::header::IntoHeaderValue;
+use bytes::Bytes;
 
 pub mod basic;
 pub mod bearer;
 
 /// Authentication challenge for `WWW-Authenticate` header.
-pub trait Challenge: IntoHeaderValue + Debug + Display + Clone + Send + Sync  {
+pub trait Challenge: IntoHeaderValue + Debug + Display + Clone + Send + Sync {
+    /// Converts the challenge into a bytes suitable for HTTP transmission.
     fn to_bytes(&self) -> Bytes;
 }