+++ src/headers/www_authenticate/header.rs
@@ -1,36 +1,18 @@
-use actix_web::{HttpMessage};
 use actix_web::error::ParseError;
 use actix_web::http::header::{Header, HeaderName, HeaderValue, IntoHeaderValue, WWW_AUTHENTICATE};
+use actix_web::HttpMessage;
 
 use super::Challenge;
 
 /// `WWW-Authenticate` header, described in [RFC 7235](https://tools.ietf.org/html/rfc7235#section-4.1)
 ///
-/// `WWW-Authenticate` header is generic over [Challenge](./trait.Challenge.html)
-///
-/// # Example
-///
-/// ```rust
-/// # extern crate actix_web;
-/// # extern crate actix_web_httpauth;
-///
-/// use actix_web::{HttpRequest, HttpResponse};
-/// use actix_web::http::StatusCode;
-/// use actix_web_httpauth::headers::www_authenticate::{WWWAuthenticate};
-/// use actix_web_httpauth::headers::www_authenticate::basic::Basic;
-///
-/// fn handler(req: HttpRequest) -> HttpResponse {
-///     let challenge = Basic {
-///         realm: Some("Restricted area".to_string()),
-///     };
-///     req.build_response(StatusCode::UNAUTHORIZED)
-///         .set(WWWAuthenticate(challenge))
-///         .finish()
-/// }
-/// ```
-pub struct WWWAuthenticate<C: Challenge>(pub C);
+/// This header is generic over [Challenge](./trait.Challenge.html) trait,
+/// see [Basic](./basic/struct.Basic.html) and [Bearer](./bearer/struct.Bearer.html)
+/// challenges for details.
+#[derive(Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Default, Clone)]
+pub struct WwwAuthenticate<C: Challenge>(pub C);
 
-impl<C: Challenge> Header for WWWAuthenticate<C> {
+impl<C: Challenge> Header for WwwAuthenticate<C> {
     fn name() -> HeaderName {
         WWW_AUTHENTICATE
     }
@@ -40,7 +22,7 @@ impl<C: Challenge> Header for WWWAuthenticate<C> {
     }
 }
 
-impl<C: Challenge> IntoHeaderValue for WWWAuthenticate<C> {
+impl<C: Challenge> IntoHeaderValue for WwwAuthenticate<C> {
     type Error = <C as IntoHeaderValue>::Error;
 
     fn try_into(self) -> Result<HeaderValue, <Self as IntoHeaderValue>::Error> {